+++ packages/ragbits-core/src/ragbits/core/vector_stores/hybrid.py
@@ -1,4 +1,3 @@
-import abc
 import asyncio
 from uuid import UUID
 
@@ -10,55 +9,7 @@ from ragbits.core.vector_stores.base import (
     VectorStoreResult,
     WhereQuery,
 )
-
-
-class HybridRetrivalStrategy(abc.ABC):
-    """
-    A class that can join vectors retrieved from different vector stores into a single list,
-    allowing for different strategies for combining results.
-    """
-
-    @abc.abstractmethod
-    def join(self, results: list[list[VectorStoreResult]]) -> list[VectorStoreResult]:
-        """
-        Joins the multiple lists of results into a single list.
-
-        Args:
-            results: The lists of results to join.
-
-        Returns:
-            The joined list of results.
-        """
-
-
-class OrderedHybridRetrivalStrategy(HybridRetrivalStrategy):
-    """
-    A class that orders the results by score and deduplicates them by choosing the first occurrence of each entry.
-    """
-
-    def __init__(self, reverse: bool = False) -> None:
-        """
-        Constructs a new OrderedHybridRetrivalStrategy instance.
-
-        Args:
-            reverse: if True orders the results in descending order by score, otherwise in ascending order.
-        """
-        self._reverse = reverse
-
-    def join(self, results: list[list[VectorStoreResult]]) -> list[VectorStoreResult]:
-        """
-        Joins the multiple lists of results into a single list.
-
-        Args:
-            results: The lists of results to join.
-
-        Returns:
-            The joined list of results.
-        """
-        all_results = [result for sublist in results for result in sublist]
-        all_results.sort(key=lambda result: result.score, reverse=self._reverse)
-
-        return list({result.entry.id: result for result in all_results}.values())
+from ragbits.core.vector_stores.hybrid_strategies import HybridRetrivalStrategy, OrderedHybridRetrivalStrategy
 
 
 class HybridSearchVectorStore(VectorStore):