+++ packages/ragbits-core/tests/unit/embeddings/test_from_config.py
@@ -12,7 +12,7 @@ def test_subclass_from_config_litellm():
         {
             "type": "ragbits.core.embeddings.litellm:LiteLLMEmbedder",
             "config": {
-                "model": "some_model",
+                "model_name": "some_model",
                 "default_options": {
                     "option1": "value1",
                     "option2": "value2",
@@ -22,7 +22,7 @@ def test_subclass_from_config_litellm():
     )
     embedder: Embedder = Embedder.subclass_from_config(config)
     assert isinstance(embedder, LiteLLMEmbedder)
-    assert embedder.model == "some_model"
+    assert embedder.model_name == "some_model"
     assert embedder.default_options == LiteLLMEmbedderOptions(
         dimensions=NOT_GIVEN,
         timeout=NOT_GIVEN,
@@ -66,7 +66,7 @@ def test_from_config_with_router():
     config = ObjectConstructionConfig(
         type="ragbits.core.embeddings.litellm:LiteLLMEmbedder",
         config={
-            "model": "text-embedding-3-small",
+            "model_name": "text-embedding-3-small",
             "api_key": "test_api_key",
             "router": [
                 {
@@ -91,7 +91,7 @@ def test_from_config_with_router():
     embedder: Embedder = Embedder.subclass_from_config(config)
     assert isinstance(embedder, LiteLLMEmbedder)
     assert embedder.api_base is None
-    assert embedder.model == "text-embedding-3-small"
+    assert embedder.model_name == "text-embedding-3-small"
     assert embedder.api_key == "test_api_key"
     assert isinstance(embedder.router, litellm.router.Router)
     assert len(embedder.router.model_list) == 2