+++ packages/ragbits-core/tests/unit/vector_stores/test_in_memory.py
@@ -4,11 +4,11 @@ import pytest
 from pydantic import computed_field
 
 from ragbits.core.embeddings.noop import NoopEmbedder
+from ragbits.core.sources.local import LocalFileSource
 from ragbits.core.vector_stores.base import EmbeddingType, VectorStoreOptions
 from ragbits.core.vector_stores.in_memory import InMemoryVectorStore
 from ragbits.document_search.documents.document import DocumentMeta, DocumentType
 from ragbits.document_search.documents.element import Element
-from ragbits.document_search.documents.sources import LocalFileSource
 
 
 class AnimalElement(Element):
@@ -173,15 +173,15 @@ async def vector_store_fixture(request: pytest.FixtureRequest) -> InMemoryVector
 
 
 @pytest.mark.parametrize(
-    ("k", "max_distance", "results"),
+    ("k", "score_threshold", "results"),
     [
         (5, None, ["spikey", "fluffy", "slimy", "spotty", "scaly"]),
         (2, None, ["spikey", "fluffy"]),
-        (5, 0.3, ["spikey", "fluffy"]),
+        (5, -0.3, ["spikey", "fluffy"]),
     ],
 )
-async def test_retrieve(store: InMemoryVectorStore, k: int, max_distance: float | None, results: list[str]) -> None:
-    query_results = await store.retrieve("query", options=VectorStoreOptions(k=k, max_distance=max_distance))
+async def test_retrieve(store: InMemoryVectorStore, k: int, score_threshold: float | None, results: list[str]) -> None:
+    query_results = await store.retrieve("query", options=VectorStoreOptions(k=k, score_threshold=score_threshold))
 
     assert len(query_results) == len(results)
     for query_result, result in zip(query_results, results, strict=True):