+++ packages/ragbits-core/CHANGELOG.md
@@ -2,7 +2,19 @@
 
 ## Unreleased
 
+- Image embeddings in PgVectorStore (#495)
+- Add PgVectorStore to vector store integration tests (#495)
+- Add new fusion strategies for the hybrid vector store: RRF and DBSF (#413)
+- move sources from ragbits-document-search to ragbits-core (#496)
+- adding connection check to Azure get_blob_service (#502)
+
+## 0.13.0 (2025-04-02)
+- Make the score in VectorStoreResult consistent (always bigger is better)
 - Add router option to LiteLLMEmbedder (#440)
+- Make LLM / Embedder APIs consistent (#463)
+- New methods in Prompt class for appending conversation history (#480)
+- Fix: make unflatten_dict symmetric to flatten_dict (#461)
+- Cost and capabilities config for custom litellm models (#481)
 
 ## 0.12.0 (2025-03-25)
 - Allow Prompt class to accept the asynchronous response_parser. Change the signature of parse_response method.
@@ -10,7 +22,6 @@
 - Fix Qdrant vector store serialization (#419)
 
 ## 0.11.0 (2025-03-25)
-
 - Add HybridSearchVectorStore which can aggregate results from multiple VectorStores (#412)
 
 ## 0.10.2 (2025-03-21)